refactor: Modify User entity path
