+++ studit/src/main/java/com/example/studit/domain/posting/Posting.java
@@ -2,7 +2,7 @@ package com.example.studit.domain.posting;
 
 import com.example.studit.domain.BaseEntity;
 import com.example.studit.domain.Comment;
-import com.example.studit.domain.User;
+import com.example.studit.domain.User.User;
 import com.example.studit.domain.enumType.Category;
 import lombok.Builder;
 import lombok.Getter;
