+++ studit/src/main/java/com/example/studit/domain/study/ParticipatedStudy.java
@@ -1,7 +1,7 @@
 package com.example.studit.domain.study;
 
 import com.example.studit.domain.BaseEntity;
-import com.example.studit.domain.User;
+import com.example.studit.domain.User.User;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
